+++ lib/solidpod.dart
@@ -172,6 +172,11 @@ export 'src/solid/api/rest_api.dart'
 export 'src/solid/api/common_permission.dart'
     show PermissionLogLiteral, getLatestLog;
 
+/// Function to revoke permission to a deleted file
+
+export 'src/solid/revoke_permission_to_deleted_file.dart'
+    show revokePermissionToDelFile;
+
 /// Read encrypted/non-encrypted files stored in an external POD
 
 export 'src/solid/read_external_pod.dart';
@@ -181,9 +186,17 @@ export 'src/solid/read_external_pod.dart';
 export 'src/solid/write_external_pod.dart';
 
 /// 20250917 gjw Extras that were required for notepod. Not yet documented.
+/// 20251103 jesscmoore In common.dart, only authUserPred is
+/// used by notepod
 
 export 'src/solid/constants/common.dart' show dataDir, profCard, authUserPred;
+
+/// Adds the permission data to a data map for a file in a user's POD
+
 export 'src/solid/get_access_lists.dart';
+
+/// Function to get resources in a user's POD
+
 export 'src/solid/get_resources.dart';
 
 /// 20250917 gjw Extras that were required for the example app! Not yet