+++ lib/src/solid/grant_permission.dart
@@ -37,6 +37,7 @@ import 'package:solidpod/src/solid/api/grant_permission_api.dart';
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants/common.dart';
 import 'package:solidpod/src/solid/constants/web_acl.dart';
+import 'package:solidpod/src/solid/models/log_entry.dart';
 import 'package:solidpod/src/solid/solid_func_call_status.dart';
 import 'package:solidpod/src/solid/utils/authdata_manager.dart';
 import 'package:solidpod/src/solid/utils/exceptions.dart';
@@ -197,13 +198,13 @@ Future<SolidFunctionCallStatus> grantPermission({
           final userWebId = await AuthDataManager.getWebId() as String;
 
           for (final recipientWebId in recipientWebIdList) {
-            final logEntryRes = createPermLogEntry(
-              permissionList,
-              resourceUrl,
-              ownerWebId,
-              'grant',
-              userWebId,
-              recipientWebId as String,
+            final LogEntry logEntryRes = createPermLogEntry(
+              permissionList: permissionList,
+              resourceUrl: resourceUrl,
+              ownerWebId: ownerWebId,
+              permissionType: 'grant',
+              granterWebId: userWebId,
+              recepientWebId: recipientWebId as String,
             );
 
             // Log file urls of the owner, granter, and receiver
@@ -217,18 +218,16 @@ Future<SolidFunctionCallStatus> grantPermission({
 
             // Run log entry insert query for the granter
             await addPermLogLine(
-              granterLogFileUrl,
-              logEntryRes[0] as String,
-              logEntryRes[1] as String,
+              logFileUrl: granterLogFileUrl,
+              logEntry: logEntryRes,
             );
 
             // If owner and the granter is not the same add another log file entry
             // for the owner
             if (ownerLogFileUrl != granterLogFileUrl) {
               await addPermLogLine(
-                ownerLogFileUrl,
-                logEntryRes[0] as String,
-                logEntryRes[1] as String,
+                logFileUrl: ownerLogFileUrl,
+                logEntry: logEntryRes,
               );
             }
 
@@ -239,9 +238,8 @@ Future<SolidFunctionCallStatus> grantPermission({
                   await getFileUrl(logFilePath, recipientWebId);
 
               await addPermLogLine(
-                receiverLogFileUrl,
-                logEntryRes[0] as String,
-                logEntryRes[1] as String,
+                logFileUrl: receiverLogFileUrl,
+                logEntry: logEntryRes,
               );
             }
           }